+++ src/Xabbo/ViewModels/Chat/ChatLogEntryViewModel.cs
@@ -2,7 +2,7 @@ namespace Xabbo.ViewModels;
 
 public abstract class ChatLogEntryViewModel : ViewModelBase
 {
-    public long EntryId { get; set; }
+    public int EntryId { get; set; }
     public DateTime Timestamp { get; } = DateTime.Now;
 
     public virtual bool IsSelectable => true;
