+++ src/Xabbo/ViewModels/Room/Bans/RoomBanViewModel.cs
@@ -1,7 +1,7 @@
 ﻿namespace Xabbo.ViewModels;
 
-public class RoomBanViewModel(long id, string name) : ViewModelBase
+public class RoomBanViewModel(int id, string name) : ViewModelBase
 {
-    public long Id { get; } = id;
+    public int Id { get; } = id;
     public string Name { get; } = name;
 }
